+++ R/esp_get_grid_ESDAC.R
@@ -135,7 +135,7 @@ esp_get_grid_ESDAC <- function(resolution = 10,
   }
 
   err_onload <- try(
-    data_sf <- sf::st_read(
+    sf::st_read(
       init_grid,
       quiet = isFALSE(verbose),
       stringsAsFactors = FALSE